refactor: implement handler pattern for command processing and cleanup config
- Extract command logic into dedicated handler classes (DownloadHandler, PutPasswordHandler)
- Create Handler interface for consistent command structure across handlers
- Refactor msg:text event to use handler pattern instead of inline logic
- Export CustomContext type from bot config for use in handlers
- Simplify .env.example by removing comments and extra configuration
- Remove unused .env.variables file
- Fix package name from pdfstudio-bot to pdf-studio-bot for consistency
- Improve code organization and maintainability with separation of concerns
