+++ .env.example
@@ -1,11 +1,2 @@
-# Configurações do Telegram Bot
 BOT_TOKEN=your_telegram_bot_token_here
-
-# Configurações do MongoDB (use sua instância externa)
 MONGODB_CONNECTION_STRING=mongodb://localhost:27017/pdf_studio
-
-# Configurações de Log
-LOG_LEVEL=info
-
-# Configurações do Node.js
-NODE_ENV=production
