--- .env.variables
@@ -1,2 +1,0 @@
-BOT_TOKEN=
-MONGODB_CONNECTION_STRING=
