+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gabrielrufino/pdfstudio-bot",
+  "name": "@gabrielrufino/pdf-studio-bot",
   "private": true,
   "version": "1.1.1",
   "main": "dist/index.js",
