+++ src/interfaces/handler.interface.ts
@@ -1,0 +1,8 @@
+import type { FilterQuery } from 'grammy'
+import type { CustomContext } from '../config/bot'
+import type { CommandEnum } from '../enumerables/command.enum'
+
+export interface Handler {
+  readonly command: CommandEnum
+  readonly events: Partial<Record<FilterQuery, (context: CustomContext) => Promise<void>>>
+}
